+++ crates/client/db/src/contract_db.rs
@@ -9,7 +9,7 @@ use std::sync::Arc;
 
 use rayon::{iter::ParallelIterator, slice::ParallelSlice};
 use rocksdb::{BoundColumnFamily, IteratorMode, ReadOptions, WriteOptions};
-use starknet_core::types::Felt;
+use starknet_types_core::felt::Felt;
 
 use crate::{
     db_block_id::{DbBlockId, DbBlockIdResolvable},